refactor: extract Pi-compatible watcher core (#2)
