+++ bin/fm-pi-compatible-runtimes
@@ -1,0 +1,2 @@
+pi
+omp
